Added all dependencies to package.json (#11)
It seems to be a better practice to keep all dependencies documented in repo and package-lock, so there will be no issues with installing them manually. Even though there are more deps that self-hosters might not use, it makes it easier to change configuration and manage some things like docker images (see #8).